implements AppsAggregator
